Define component within module
Instead of importing all of the modules into the `index.js` file and
defining them there, it's handy to instead define them right after the
class has been created. This way, as soon as a component is imported,
the Web Component is registered.

For the time being, all classes will be imported into the `index.js`
file to be registered. However, this might be updated in the future to
allow a user to specify whether a component is used from HTML or _only_
from another component. This way, "internal" components would be
imported where they are needed, allowing them to be more easily
tree-shaken from a build if not in use.