Replace Uglify with Babili
Since we're using `babel-preset-env` for determining what code needs to
be transpiled by Babel, we may end up in a position (depending on
browser needs) where we're actually going to shop ES6 code to clients.
If that's the case, Uglify will choke on the output, but Babili will
work fine.

An initial comparison shows that the Babili output is slightly smaller
for the initial code produced when running the Skate app generator.